working on cell data

+++ app/src/main/java/com/cellwatch/data/local/model/CellEntity.kt
@@ -11,8 +11,8 @@ data class CellEntity(
     var id: String = UUID.randomUUID().toString(),
     val timestamp: Instant? = null,
 
-    val cellId: Long? = null,
-    val physicalCellId: Long? = null,
+    val cellId: Int? = null,
+    val physicalCellId: Int? = null,
     val cellConnection: Int? = null,
     val networkGeneration: String? = null,
     val networkSubtype: String? = null,
