+++ app/src/main/java/com/cellwatch/data/model/Cell.kt
@@ -8,8 +8,8 @@ import java.util.UUID
 data class Cell(
     var id: String = UUID.randomUUID().toString(),
     val timestamp: Instant? = null,
-    val cellId: Long? = null,
-    val physicalCellId: Long? = null,
+    val cellId: Int? = null,
+    val physicalCellId: Int? = null,
     val cellConnection: Int? = null,
     val networkGeneration: String? = null,
     val networkSubtype: String? = null,
